+++ cate/ds/esa_cci_odp.py
@@ -313,13 +313,11 @@ class EsaCciOdpDataStore(DataStore):
                         data_source.update_file_list()
                         child_monitor.progress(work=1)
 
-    def query(self, name: str = None, monitor: Monitor = Monitor.NONE) -> Sequence['DataSource']:
+    def query(self, id: str = None, query_expr: str = None, monitor: Monitor = Monitor.NONE) -> Sequence['DataSource']:
         self._init_data_sources()
-        if name:
-            result = [data_source for data_source in self._data_sources if data_source.matches_filter(name)]
-        else:
-            result = self._data_sources
-        return result
+        if id or query_expr:
+            return [ds for ds in self._data_sources if ds.matches(id=id, query_expr=query_expr)]
+        return self._data_sources
 
     def _repr_html_(self) -> str:
         self._init_data_sources()
@@ -574,7 +572,7 @@ class EsaCciOdpDataSource(DataSource):
                      time_range: TimeRangeLike.TYPE,
                      monitor: Monitor = Monitor.NONE) -> bool:
 
-        data_sources = find_data_sources(None, local_id)  # type: Sequence['DataSource']
+        data_sources = find_data_sources(None, id=local_id)  # type: Sequence['DataSource']
         data_source = next((ds for ds in data_sources if isinstance(ds, LocalDataSource) and
                             ds.name == local_id), None)  # type: LocalDataSource
         if not data_source: